refactor: divide logistic regression to other package
